sfst+ng-index: widen the batch length check to u64; state the per-facet capture rule precisely
The StreamBatch integrity compare multiplied a u32 total by 4 in
usize — on the 32-bit static builds (armv6/armv7) a crafted chunk
could wrap the multiply and slip past the very check meant to catch
it. Compared in u64 the multiply cannot wrap anywhere, and a passing
check also proves every row() slice bound fits usize. The seal-capture
comment now states the exact per-facet first-entry rule instead of
overclaiming uniformity (the nuances are unreachable from the
flattener's typed output).

+++ src/crates/sfst/src/schema.rs
@@ -856,8 +856,10 @@ impl StreamBatch {
             self.row_offsets.push(acc);
         }
         // Every KvId is a fixed 4 bytes, so the buffer must hold exactly
-        // `total_ids * 4`.
-        self.kv_bytes.len() == acc as usize * 4
+        // `total_ids * 4` — compared in u64 so a 32-bit `usize` (the
+        // armv6/armv7 static builds) cannot wrap the multiply; a check
+        // that passes also proves every `row()` slice bound fits usize.
+        self.kv_bytes.len() as u64 == u64::from(acc) * 4
     }
 
     /// Number of log rows in this batch.
